separate ressource API
