+++ recoco/apps/resources/views.py
@@ -33,7 +33,7 @@ from rest_framework import permissions, viewsets
 from reversion_compare.views import HistoryCompareDetailView
 
 from . import models
-from .serializers import ResourceSerializer
+
 
 ########################################################################
 # Searching resources
@@ -476,23 +476,6 @@ def delete_bookmark(request, resource_id=None):
             pass
     next_url = reverse("resources-resource-detail", args=[resource_id])
     return redirect(next_url)
-
-
-########################################################################
-# REST API
-########################################################################
-class ResourceViewSet(viewsets.ModelViewSet):
-    """
-    API endpoint that allows resources to be listed or edited
-    """
-
-    def get_queryset(self):
-        return models.Resource.on_site.exclude(status=models.Resource.DRAFT).order_by(
-            "-created_on", "-updated_on"
-        )
-
-    serializer_class = ResourceSerializer
-    permission_classes = [permissions.IsAuthenticatedOrReadOnly]
 
 
 # eof
